pyproject.toml
[tool.poetry]
name = "message-integrity"
version = "0.1.0"
description = ""
authors = ["AlexNistyuk <alex.nistyuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
channels = "4.0.0"
daphne = "4.0.0"
channels-redis = "4.1.0"
djangorestframework = "^3.15.2"
aioimaplib = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
